requirements.txt
discord.py
pyngrok